+++ src/docker/docker.service.ts
@@ -1,4 +1,4 @@
-import * as Docker from 'dockerode';
+import Dockerode, * as Docker from 'dockerode';
 
 import { Injectable } from '@nestjs/common';
 
@@ -16,13 +16,45 @@ export class DockerService {
     return this.docker.ping();
   }
 
-  async run(image, options?: RunOptions) {
-    await this.docker.pull(image);
-    return this.docker.run(image, options?.command, null, {
-      Env: options?.env,
-      Cmd: options?.command,
-      Labels: options?.labels,
-      HostConfig: { NetworkMode: 'host' },
+  async pull(image: string) {
+    return new Promise((resolve, reject) => {
+      this.docker.pull(image, (err, stream) => {
+        if (err) {
+          return reject(err);
+        }
+        this.docker.modem.followProgress(stream, onFinished);
+        function onFinished(err, output) {
+          if (err) {
+            return reject(err);
+          }
+          return resolve(output);
+        }
+      });
+    });
+  }
+
+  async run(image, options?: RunOptions): Promise<Dockerode.Container> {
+    await this.pull(image);
+    return new Promise((resolve, reject) => {
+      this.docker.createContainer(
+        {
+          Image: image,
+          Tty: true,
+          Cmd: ['yarn', 'start:prod'],
+          Labels: options?.labels,
+          HostConfig: { NetworkMode: 'host' },
+          Env: options?.env,
+        },
+        (err, container) => {
+          container.start({}, (err, data) => {
+            if (err) {
+              reject(err);
+              return;
+            }
+            resolve(container);
+          });
+        },
+      );
     });
   }
 }