+++ src/games/commands/handlers/create-player.handler.ts
@@ -1,15 +1,34 @@
 import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
 
+import { ConfigService } from '@nestjs/config';
 import { CreatePlayerCommand } from '../create-player.command';
 import { DockerService } from 'src/docker/docker.service';
+import { GameRepository } from 'src/games/repository/game.repository';
 
 @CommandHandler(CreatePlayerCommand)
 export class CreatePlayerHandler
   implements ICommandHandler<CreatePlayerCommand>
 {
-  constructor(private readonly service: DockerService) {}
+  constructor(
+    private readonly config: ConfigService,
+    private readonly service: DockerService,
+    private readonly repository: GameRepository,
+  ) {}
 
   async execute(command: CreatePlayerCommand) {
-    this.service.run(command.image);
+    const player = command.image.replace(/[^a-zA-Z0-9]/, '-');
+    const container = await this.service.run(command.image, {
+      command: ['yarn', 'prod:start'],
+      env: [
+        'RABBITMQ_URL=amqp://localhost:5672',
+        'EXCHANGE_NAME=games.exchange',
+        `QUEUE_NAME=player.${player}`,
+        `ROUTING_KEY=${player}`,
+      ],
+      labels: { yifan: 'player' },
+    });
+    this.repository.update(command.game, {
+      [`${command.player}.container`]: container.id,
+    });
   }
 }