+++ src/games/queries/get-game.query.ts
@@ -1,0 +1,3 @@
+export class GetGameQuery {
+  constructor(public readonly id: string) {}
+}