+++ src/games/sagas/games.sagas.ts
@@ -1,10 +1,12 @@
 import { ICommand, Saga, ofType } from '@nestjs/cqrs';
 import { Observable, of } from 'rxjs';
+import { map, mergeMap } from 'rxjs/operators';
 
 import { CreatePlayerCommand } from '../commands/create-player.command';
 import { GameCreatedEvent } from '../events/game-created.event';
 import { Injectable } from '@nestjs/common';
-import { mergeMap } from 'rxjs/operators';
+import { MovePlayedEvent } from '../events/move-played.event';
+import { PlayMoveCommand } from '../commands/play-move.command';
 
 @Injectable()
 export class GameSagas {
@@ -14,10 +16,19 @@ export class GameSagas {
       ofType(GameCreatedEvent),
       mergeMap((event) =>
         of(
-          new CreatePlayerCommand(event.gameId, event.whiteImage),
-          new CreatePlayerCommand(event.gameId, event.blackImage),
+          new CreatePlayerCommand(event.gameId, 'white', event.whiteImage),
+          new CreatePlayerCommand(event.gameId, 'black', event.blackImage),
+          new PlayMoveCommand(event.gameId, 'white'),
         ),
       ),
+    );
+  };
+
+  @Saga()
+  movePlayed = (events$: Observable<any>): Observable<ICommand> => {
+    return events$.pipe(
+      ofType(MovePlayedEvent),
+      map((event) => new PlayMoveCommand(event.game, event.player === 'white'? 'black' : 'white'),),
     );
   };
 }